Add 90 degree symmetry mode to autorotate, some averaging when setpoint far away, tune hero chsasis constants

+++ aruw-mcb-project/src/aruwsrc/control/chassis/beyblade_command.hpp
@@ -73,7 +73,7 @@ private:
     ChassisSubsystem* chassis;
     const tap::control::turret::TurretSubsystemInterface* turret;
 
-    modm::interpolation::Linear<modm::Pair<float, float>> rotationalSpeedLinearInterpolator;
+    modm::interpolation::Linear<modm::Pair<int, float>> rotationalSpeedLinearInterpolator;
 
     float getRotationTarget() const;
 };  // class BeybladeCommand
